Revert "Shape int->double"
This reverts commit 881f11fa85d3fd5c4d974025b6a1d311970a37f5.

+++ sayaya-ui/src/net/sayaya/ui/regacy/graph/AxisContinuous.java
@@ -12,8 +12,8 @@ public class AxisContinuous<T extends Number> extends Axis<T> implements HasStro
 	private NumberFormat nf = NumberFormat.getDecimalFormat();
 	private String borderColor;
 	private double borderWidth;
-	private double width;
-	private double height;
+	private int width;
+	private int height;
 	public AxisContinuous(AXIS_DIRECTION direction) {
 		super(direction);
 	}
@@ -84,8 +84,8 @@ public class AxisContinuous<T extends Number> extends Axis<T> implements HasStro
 		if(isInverted()) context.scale(1.0, -1.0);
 		Double p = min.doubleValue();
 		double step = this.step.doubleValue();
-		double stepSize = 10;
-		double textLoc = 20;
+		int stepSize = 10;
+		int textLoc = 20;
 		while(p.compareTo(max.doubleValue()) <= 0) {
 			context.beginPath();
 			context.moveTo((p+translate)*scaleX, 0);
@@ -122,20 +122,20 @@ public class AxisContinuous<T extends Number> extends Axis<T> implements HasStro
 			return getY()-(data.doubleValue() + translate) * scaleX;
 		else return getX() + (data.doubleValue() + translate) * scaleX;
 	}
-	public double getWidth() {
+	public int getWidth() {
 		return width;
 	}
 
-	public AxisContinuous<T> setWidth(double width) {
+	public AxisContinuous<T> setWidth(int width) {
 		this.width = width;
 		return this;
 	}
 
-	public double getHeight() {
+	public int getHeight() {
 		return height;
 	}
 
-	public AxisContinuous<T> setHeight(double height) {
+	public AxisContinuous<T> setHeight(int height) {
 		this.height = height;
 		return this;
 	}
